+++ pkg/merge/linstorsatelliteconfiguration_test.go
@@ -27,9 +27,9 @@ var (
 				{Name: "prop1", Value: "config1"},
 				{Name: "prop2", Value: "config1"},
 			},
-			InternalTLS: &piraeusv1.TLSConfig{
+			InternalTLS: &piraeusv1.TLSConfigWithHandshakeDaemon{TLSConfig: piraeusv1.TLSConfig{
 				SecretName: "config1",
-			},
+			}},
 		},
 	}
 	Config2 = piraeusv1.LinstorSatelliteConfiguration{
@@ -66,9 +66,9 @@ var (
 				{Name: "prop2", Value: "config3"},
 				{Name: "prop3", Value: "config3"},
 			},
-			InternalTLS: &piraeusv1.TLSConfig{
+			InternalTLS: &piraeusv1.TLSConfigWithHandshakeDaemon{TLSConfig: piraeusv1.TLSConfig{
 				SecretName: "config3",
-			},
+			}},
 		},
 	}
 )
@@ -118,9 +118,9 @@ func TestMergeSatelliteConfigurations(t *testing.T) {
 						{Name: "prop2", Value: "config3"},
 						{Name: "prop3", Value: "config3"},
 					},
-					InternalTLS: &piraeusv1.TLSConfig{
+					InternalTLS: &piraeusv1.TLSConfigWithHandshakeDaemon{TLSConfig: piraeusv1.TLSConfig{
 						SecretName: "config3",
-					},
+					}},
 				},
 			},
 		},